Test file split
