+++ mrbgem.rake
@@ -7,4 +7,5 @@ MRuby::Gem::Specification.new('mruby-ipvs') do |spec|
                 Dir.glob("#{spec.dir}/ipvsadm/libipvs/*.c")).map do |f|
     f.relative_path_from(spec.dir).pathmap("#{build_dir}/%X.o")
   end
+  spec.add_test_dependency 'mruby-require'
 end
